Update azauth, token gets, filefuncs, pricing sheet generator
